temp upload of customized cv6
temporarily uploading my personalized cv6 theme for use in quickcss @import, since discord is being stupid and this is apparently the current fix to stop flickering/refreshing. 

i'm not eager to have to push an update to this repo every time i make edits/class fixes to the theme, so hopefully normal local theme usage will be fixed eventually.